#232, reduce mysql restarts by not forcing every service to restart
The pull and start images task no longer forces containers to restart, which is needed when the environment variable is "dev" to force the demo data to be reloaded.  Now the fpm container will force-restart if dev is set for the profile, as well as any container is re-created if the image changes.  This is an incremental step.