#232, move pcmt env to play from task
PCMT_VER in particular when set at the task level didn't apply to both docker tasks, resulting in the env being unset in the second task.  By moving the env setting to the play, we can ensure it's available for all tasks and simplify how we think of the environment.